backend/requirements.txt
# 数据采集
akshare>=1.12.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 可视化
matplotlib>=3.7.0

# 定时任务
schedule>=1.2.0

# 数据库（资产配置模块）
sqlalchemy>=2.0.0
pymysql>=1.1.0

# API 服务
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
httpx>=0.27.0
python-dotenv>=1.0.1

# 日志（Python内置，无需安装）
